chore: refactor document_converter.
