+++ src/langchain_google_firestore/__init__.py
@@ -13,6 +13,5 @@
 # limitations under the License.
 
 from langchain_google_firestore.document_loader import FirestoreLoader, FirestoreSaver
-from langchain_google_firestore.utility.document_converter import DocumentConverter
 
 __all__ = ["FirestoreLoader", "FirestoreSaver"]
